requirements.txt
googlemaps==4.10.0
langchain==0.0.263
openai==0.27.8
folium==0.14.0
branca==0.6.0
python-dotenv==1.0.0
black==23.7.0
google-generativeai==0.1.0rc1
gradio==3.42.0
leafmap==0.23.4
geopandas==0.13.2